Refactor, Bugfix (apps/pet-shop): Correct the imports, rename `PSProductsStorage`
