+++ apps/pet-shop/src/app/integrator/data/ps-products/products-local.service.spec.ts
@@ -1,11 +1,13 @@
-import { TestBed } from '@angular/core/testing';
-import { ProductsLocalService } from './products-local.service';
+import { TestBed } from '@angular/core/testing'
+import { ProductsLocalService } from './products-local.service'
+
+// TODO: Explore the tests, refactor, {migrate to `vitest`?}.
 
 describe('ProductsLocalService', (): void => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  beforeEach((): TestBed => TestBed.configureTestingModule({}))
 
   it('should be created', (): void => {
-    const service: ProductsLocalService = TestBed.get(ProductsLocalService);
-    expect(service).toBeTruthy();
-  });
-});
+    const service: ProductsLocalService = TestBed.get(ProductsLocalService)
+    expect(service).toBeTruthy()
+  })
+})
