+++ apps/pet-shop/src/app/integrator/data/ps-products/products-local.service.ts
@@ -3,8 +3,8 @@ import { Injectable } from '@angular/core'
 import { type Observable, of } from 'rxjs'
 
 // # Internal modules
-import type { PSProductsService } from '../../../../integrator/ui/ps-products/products-service.type'
-import type { PSProductTableItem } from '../../../../integrator/ui/ps-products/ps-products.type'
+import type { PSProductsService } from '../../ui/ps-products/products-service.type'
+import type { PSProductTableItem } from '../../ui/ps-products/ps-products.type'
 import productsData from './products.data'
 
 @Injectable({
